change to_d for to_f for Heroku

+++ app/models/project.rb
@@ -23,7 +23,7 @@ class Project < ActiveRecord::Base
   def note_for(category)
     values = notes.where(category_id: category.id).pluck(:value)
     return nil if values.empty?
-    (values.sum / values.count.to_d).round(1)
+    (values.sum / values.count.to_f).round(1)
   end
 
   def to_param
